Changed the dependency from Gson to Jackson.
The reasoning behind this is that Jackson provides some powerful tools to JSON handling while being one of the best performing libraries at it; I see no reason to use Gson instead of Jackson because if jNarwhal is ever used on a mobile platform (i.e. Android) resources and performance are going to be a critical point to the app that using the wrapper.